added support for transactions together with support for consumer.Receive(timeout)
git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/activemq/trunk@381643 13f79535-47bb-0310-9956-ffa450edef68

+++ openwire-dotnet/src/OpenWire.Client/IMessageConsumer.cs
@@ -35,7 +35,7 @@ namespace OpenWire.Client
         /// <summary>
         /// If a message is available within the timeout duration it is returned otherwise this method returns null
         /// </summary>
-        IMessage Receive(long timeout);
+        IMessage Receive(int timeout);
         
         /// <summary>
         /// If a message is available immediately it is returned otherwise this method returns null
